+++ frontend/src/routes/__root.tsx
@@ -16,9 +16,12 @@ export const Route = createRootRoute({
               <span className="font-semibold text-lg">Car Rental</span>
             </Link>
             <div className="flex items-center gap-3">
+              <Link to="/browse">
+                <Button variant="ghost" size="sm">Browse</Button>
+              </Link>    
               <Link to="/auth">
                 <Button variant="ghost" size="sm">Log in</Button>
-              </Link>
+              </Link>           
               <Link to="/auth">
                 <Button size="sm">Sign up</Button>
               </Link>
